Adding docstrings
