+++ pyreal/realapp/realapp.py
@@ -43,8 +43,7 @@ def _format_feature_importance_output(explanation):
             Pyreal Explanation object to parse
 
     Returns:
-        One dataframe per id, with each row representing a feature, and four columns:
-            Feature Name    Feature Value   Contribution    Average/Mode
+        DataFrame with a Feature Name column and an Importance column
     """
     importances = explanation.get()
     return pd.DataFrame({"Feature Name": importances.columns, "Importance": importances.squeeze()})
@@ -144,6 +143,16 @@ class RealApp:
         # {"explanation_type": {"algorithm":Explainer} }
 
     def _make_base_explainer(self, model):
+        """
+        Make a base explainer for model.
+
+        Args:
+            model (model object):
+                The model to be explained by this explainer
+        Returns:
+            Explainer
+                The explainer
+        """
         return Explainer(
             model,
             self.X_train_orig,
@@ -153,18 +162,116 @@ class RealApp:
         )
 
     def _explainer_exists(self, explanation_type, algorithm):
+        """
+        Check if the requested explainer exists
+
+        Args:
+            explanation_type (string):
+                Code for explanation_type
+            algorithm (string):
+                Name of algorithm
+
+        Returns:
+            Boolean
+                True if the specified explainer exists, False otherwise
+        """
         if explanation_type in self.explainers:
             if algorithm in self.explainers[explanation_type]:
                 return True
         return False
 
     def _add_explainer(self, explanation_type, algorithm, explainer):
+        """
+        Add the specified explainer to this RealApp
+
+        Args:
+            explanation_type (string):
+                Code for explanation_type
+            algorithm (string):
+                Name of algorithm
+            explainer (Explainer):
+                Explainer to add
+        """
         if explanation_type not in self.explainers:
             self.explainers[explanation_type] = {}
         self.explainers[explanation_type][algorithm] = explainer
 
     def _get_explainer(self, explanation_type, algorithm):
+        """
+        Get the requested explainer
+
+        Args:
+            explanation_type (string):
+                Code for explanation_type
+            algorithm (string):
+                Name of algorithm
+
+        Returns:
+            Explainer
+                The requested explainer
+        """
         return self.explainers[explanation_type][algorithm]
+
+    def _produce_explanation_helper(
+        self,
+        explanation_type_code,
+        algorithm,
+        prepare_explainer_func,
+        format_output_func,
+        x_orig=None,
+        model_id=None,
+        id_column_name=None,
+        force_refit=False,
+        **kwargs
+    ):
+        """
+        Produce an explanation from a specified Explainer
+
+        Args:
+            explanation_type (string):
+                Code for explanation_type
+            algorithm (string):
+                Name of algorithm
+            prepare_explainer_func (function):
+                Function that initializes and fits the appropriate explainer
+            format_output_func (function):
+                Function that formats Explanation objects into the appropriate output format
+            x_orig (DataFrame):
+                Data to explain, required for local explanations
+            model_id (string or int):
+                ID of model to explain
+            id_column_name (string or int):
+                Name of column that contains item ids in input data
+            force_refit (Boolean):
+                If True, initialize and fit a new explainer even if the appropriate explainer
+                already exists
+            **kwargs:
+                Additional explainer fit parameters
+
+        Returns:
+            Type varies by explanation type
+                The explanation
+        """
+        if model_id is None:
+            model_id = self.active_model_id
+
+        if self._explainer_exists(explanation_type_code, algorithm) and not force_refit:
+            explainer = self._get_explainer(explanation_type_code, algorithm)
+        else:
+            explainer = prepare_explainer_func(model_id=model_id, algorithm=algorithm, **kwargs)
+
+        if x_orig is not None:
+            if id_column_name is not None:
+                ids = x_orig[id_column_name]
+                x_orig = x_orig.drop(columns=id_column_name)
+            else:
+                ids = x_orig.index
+
+            explanation = explainer.produce(x_orig)
+            return format_output_func(explanation, ids)
+        else:
+            explanation = explainer.produce()
+            return format_output_func(explanation)
 
     def add_model(self, model, model_id=None):
         """
@@ -228,40 +335,21 @@ class RealApp:
 
         return self.base_explainers[model_id].model_predict(x)
 
-    def _produce_explanation_helper(
-        self,
-        algorithm,
-        explanation_type_code,
-        prepare_explainer_func,
-        format_output_func,
-        x_orig=None,
-        model_id=None,
-        id_column_name=None,
-        force_refit=False,
-        **kwargs
-    ):
-        if model_id is None:
-            model_id = self.active_model_id
-
-        if self._explainer_exists(explanation_type_code, algorithm) and not force_refit:
-            explainer = self._get_explainer(explanation_type_code, algorithm)
-        else:
-            explainer = prepare_explainer_func(model_id=model_id, algorithm=algorithm, **kwargs)
-
-        if x_orig is not None:
-            if id_column_name is not None:
-                ids = x_orig[id_column_name]
-                x_orig = x_orig.drop(columns=id_column_name)
-            else:
-                ids = x_orig.index
-
-            explanation = explainer.produce(x_orig)
-            return format_output_func(explanation, ids)
-        else:
-            explanation = explainer.produce()
-            return format_output_func(explanation)
-
     def prepare_local_feature_contributions(self, model_id=None, algorithm=None, shap_type=None):
+        """
+        Initialize and fit a local feature contribution explainer
+
+        Args:
+            model_id (int or string):
+                Model id to explain
+            algorithm (string):
+                LFC algorithm to use
+            shap_type (string):
+                If algorithm is "shap", type of shap to use
+
+        Returns:
+            A fit LocalFeatureContribution explainer
+        """
         if algorithm is None:
             algorithm = "shap"
 
@@ -285,12 +373,34 @@ class RealApp:
         shap_type=None,
         force_refit=False,
     ):
+        """
+        Produce a LocalFeatureContributions explainer
+
+        Args:
+            x_orig (DataFrame):
+                Input to explain
+            model_id (string or int):
+                ID of model to explain
+            algorithm (string):
+                Name of algorithm
+            id_column_name (string or int):
+                Name of column that contains item ids in input data
+            shap_type (string):
+                If algorithm="shap", type of SHAP explainer to use
+            force_refit (Boolean):
+                If True, initialize and fit a new explainer even if the appropriate explainer
+                already exists
+
+        Returns:
+            One dataframe per id, with each row representing a feature, and four columns:
+            Feature Name    Feature Value   Contribution    Average/Mode
+        """
         if algorithm is None:
             algorithm = "shap"
 
         return self._produce_explanation_helper(
-            algorithm,
             "lfc",
+            algorithm,
             self.prepare_local_feature_contributions,
             _format_feature_contribution_output,
             x_orig=x_orig,
@@ -301,6 +411,20 @@ class RealApp:
         )
 
     def prepare_global_feature_importance(self, model_id=None, algorithm=None, shap_type=None):
+        """
+        Initialize and fit a global feature importance explainer
+
+        Args:
+            model_id (int or string):
+                Model id to explain
+            algorithm (string):
+                GFI algorithm to use
+            shap_type (string):
+                If algorithm is "shap", type of shap to use
+
+        Returns:
+            A fit GlobalFeatureImportance explainer
+        """
         if algorithm is None:
             algorithm = "shap"
 
@@ -322,12 +446,29 @@ class RealApp:
         shap_type=None,
         force_refit=False,
     ):
+        """
+        Produce a GlobalFeatureImportance explainer
+
+        Args:
+            model_id (string or int):
+                ID of model to explain
+            algorithm (string):
+                Name of algorithm
+            shap_type (string):
+                If algorithm="shap", type of SHAP explainer to use
+            force_refit (Boolean):
+                If True, initialize and fit a new explainer even if the appropriate explainer
+                already exists
+
+        Returns:
+            DataFrame with a Feature Name column and an Importance column
+        """
         if algorithm is None:
             algorithm = "shap"
 
         return self._produce_explanation_helper(
-            algorithm,
             "gfi",
+            algorithm,
             self.prepare_global_feature_importance,
             _format_feature_importance_output,
             model_id=model_id,
